MIRParser: Avoid reading uninitialized memory on generic vregs
If a vreg's bank is specified in the registers block and one of its
defs or uses also specifies the bank, we end up checking that the
RegBank is equal to diagnose conflicting banks. The problem comes up
for generic vregs, where we weren't fully initializing the VRegInfo
when parsing the registers block, so we'd end up comparing a null
pointer to uninitialized memory.

This fixes a non-deterministic failure when round tripping through MIR
with generic vregs.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@318543 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 9bde5fb8f8ea44d6ef0f2edb4f54cbff7a0aee53)