AArch64: Fix emergency spillslot being out of reach for large callframes
This is a backport of https://reviews.llvm.org/D40876 which turns out to
fix rdar://35116491 and rdar://33229629

Changes compared to D40876:
- Removed mir file loading enhancements (not essential to fix the bug)
- Only change useFPForScavengingIndex() behavior when oversized
  callframes are present to minimize codegen changes compared to
  previous version.
- Removed the testcase as it needs fixes from https://reviews.llvm.org/D42006
  to pass expensive checks enabled bots (that's an unrelated bugfix I
  did not want to backport that late in the process).

Large callframes (calls with several hundreds or thousands or
parameters) could lead to situations in which the emergency spillslot is
out of range to be addressed relative to the stack pointer.
This commit forces the use of a frame pointer in the presence of large
callframes.

This commit does several things:
- Compute max callframe size at the end of instruction selection.
- Let TargetFrameLowering::hasFP() return true if there exists a
  callframe > 255 bytes.
- Place the emergency spillslot close to FP if a large callframe exists.