[DAGCombiner] Set the right SDLoc on a newly-created sextload (6/N)
This teaches tryToFoldExtOfLoad to set the right location on a
newly-created extload. With that in place, the logic for performing a
certain ([s|z]ext (load ...)) combine becomes identical for sexts and
zexts, and we can get rid of one copy of the logic.

The test case churn is due to dependencies on IROrders inherited from
the wrong SDLoc.

Part of: llvm.org/PR37262

Differential Revision: https://reviews.llvm.org/D46158

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@332118 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit bdb4eca65ea5639cbce5e5c59fa8c7ebab44ca8a)

 Conflicts:
	lib/CodeGen/SelectionDAG/DAGCombiner.cpp
	test/CodeGen/AArch64/arm64-ldp-cluster.ll
	test/CodeGen/X86/avx512-insert-extract.ll
	test/CodeGen/X86/known-signbits-vector.ll
	test/CodeGen/X86/vector-shuffle-variable-128.ll
	test/CodeGen/X86/vector-shuffle-variable-256.ll