[dwarfdump] Make -c and -p work together
When requesting to dump both the parent chain and children, we used to
print the DIE more than once because we propagated the dump options to
the parent without clearing the respective flags. This commit fixes this
oversight and adds a test.

rdar://39415292

Differential revision: https://reviews.llvm.org/D47263

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@333350 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 98ac263943471aa03c28bc78ee05778e3d3f49fd)