[DWARF Verifier] Add helper function to dump DIEs. [NFC]
It's pretty common for the verifier to dump the relevant DIE when it
finds an issue. This tends to be relatively verbose and error prone
because we have to pass the DIDumpOptions to the DIE's dump method. This
patch adds a helper function to the verifier to make this easier.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@342526 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit aef051374a9f28bea986a29da88868a91db1d893)