[test] Disable a bitcode test for AppleClang
This test just checks that the upstream llvm.org compiler can load
bitcode produced by an old version of AppleClang. The upstream compiler
discards the debug info from the test file because of a version
mismatch. However, the Apple compiler does not. Disable this test for
Apple compilers (there already are dedicated debug info bitcode
compatibility tests for it).

rdar://44645820